chore: address comments, fix linter

+++ examples-standalone/coffee-warehouse/src/app/components/profile/profile.component.ts
@@ -1,5 +1,4 @@
-import { Component } from '@angular/core';
-import { Router } from '@angular/router';
+import { AfterViewInit, Component } from '@angular/core';
 import { Validators, FormGroup, FormControl } from '@angular/forms';
 
 import { countries } from 'src/app/resources/countries';
@@ -8,22 +7,23 @@ import { FormModel } from 'src/app/models/form.model';
 import { SelectEvent, FileRestrictions } from '@progress/kendo-angular-upload';
 import { CustomMessagesService } from 'src/app/services/custom-messages.service';
 import { MessageService } from '@progress/kendo-angular-l10n';
+import { NotificationService } from '@progress/kendo-angular-notification';
 
 @Component({
-    selector: 'profile-component',
+    selector: 'app-profile-component',
     templateUrl: './profile.component.html'
 })
-export class ProfileComponent {
+export class ProfileComponent implements AfterViewInit {
     public formGroup: FormGroup;
     public countries = countries;
-    public phoneNumberMask: string = '(+9) 0000-000-00-00';
+    public phoneNumberMask = '(+9) 0000-000-00-00';
     public fileRestrictions: FileRestrictions = {
         allowedExtensions: ['.png', '.jpeg', '.jpg']
     };
     public avatars: NodeList;
 
     public formValue: FormModel | null = {
-        avatar: null,
+        avatar: [''],
         firstName: 'Peter',
         lastName: 'Douglas',
         email: 'peter.douglas@progress.com',
@@ -36,9 +36,9 @@ export class ProfileComponent {
 
     public customMsgService: CustomMessagesService;
 
-    constructor(public router: Router, public msgService: MessageService) {
+    constructor(public msgService: MessageService, private notificationService: NotificationService) {
         this.setFormValues();
-        this.customMsgService = <CustomMessagesService>this.msgService;
+        this.customMsgService = this.msgService as CustomMessagesService;
     }
 
     ngAfterViewInit() {
@@ -52,7 +52,7 @@ export class ProfileComponent {
         }
 
         this.formGroup = new FormGroup({
-            avatar: new FormControl(),
+            avatar: new FormControl(this.formValue.avatar, [Validators.required]),
             firstName: new FormControl(this.formValue.firstName, [Validators.required]),
             lastName: new FormControl(this.formValue.lastName, [Validators.required]),
             email: new FormControl(this.formValue.email, [Validators.required, Validators.email]),
@@ -78,19 +78,32 @@ export class ProfileComponent {
         this.formGroup.markAllAsTouched();
         const formValues = JSON.stringify(this.formGroup.value);
         localStorage.setItem('form', formValues);
-        this.router.navigate(['/'])
+
+        this.formGroup.markAsPristine();
+
+        this.notificationService.show({
+            content: 'Profile changes have been saved.',
+            animation: { type: 'slide', duration: 500 },
+            position: { horizontal: 'center', vertical: 'bottom' },
+            type: { style: 'success', icon: true },
+            hideAfter: 2000
+        });
     }
 
     public cancelChanges(): void {
         this.setFormValues();
     }
 
+    public isFileAllowed(file): boolean {
+        return this.fileRestrictions.allowedExtensions.includes(file.extension);
+    }
+
     public selectAvatar(ev: SelectEvent): void {
         const avatars = this.avatars;
-        let reader = new FileReader();
+        const  reader = new FileReader();
         const file = ev.files[0];
-        if (file.rawFile) {
-            reader.onloadend = function () {
+        if (file.rawFile && this.isFileAllowed(file)) {
+            reader.onloadend = function() {
                 avatars.forEach((avatar: HTMLElement) => {
                     avatar.style['background-image'] = `url("${this.result}")`;
                     localStorage.setItem('avatar', this.result.toString());
